refactor(frontend): move ELEMENT_COLORS + RechunkOptions to shared/ (#audit-05, #audit-07)
ELEMENT_COLORS was duplicated in 4 files and RechunkOptions lived inside the document feature while consumed by chunks. Both are now single-sourced under shared/ (elementColors.ts moved via git mv; RechunkOptions moved to shared/types.ts), and every importer repointed. This also removes several cross-feature imports flagged by the decoupling audit. Note: StructureViewer's legend now shows title+code too, since it previously used a partial local palette.
