+++ frontend/src/features/analysis/ui/ResultTabs.vue
@@ -195,18 +195,7 @@ import MarkdownViewer from './MarkdownViewer.vue'
 import ImageGallery from './ImageGallery.vue'
 import { useI18n } from '../../../shared/i18n'
 import type { PageElement } from '../../../shared/types'
-
-const ELEMENT_COLORS: Record<string, string> = {
-  title: '#EF4444',
-  section_header: '#F97316',
-  text: '#3B82F6',
-  table: '#8B5CF6',
-  picture: '#22C55E',
-  list: '#06B6D4',
-  formula: '#EC4899',
-  code: '#14B8A6',
-  caption: '#EAB308',
-}
+import { ELEMENT_COLORS } from '@/shared/elementColors'
 
 const props = defineProps({
   currentPage: { type: Number, default: 1 },
