+++ frontend/src/features/chunks/store.ts
@@ -1,7 +1,8 @@
 import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 import type { DocChunk, ChunkDiff } from '../../shared/types'
-import { rechunkDocument, type RechunkOptions } from '../document/api'
+import type { RechunkOptions } from '@/shared/types'
+import { rechunkDocument } from '../document/api'
 import * as api from './api'
 
 export const useChunksStore = defineStore('chunks', () => {
